+++ src/interfaces/IChart.ts
@@ -1,8 +1,10 @@
 import {IChartData, IWidgetVariables} from ".";
 
+/**
+ * NOTE: Здесь описываются только методы, которые будут доступны НАРУЖУ из библиотеки, и ничего более!
+ */
 export interface IChart {
     destroy(): void;                        // Удаление виджета
     getVariables(): IWidgetVariables;       // Возвращает список переменных, доступных для прослушивания
-    run(data: IChartData): void;            // Запуск виджета
     redraw(): Promise<void>;                // Перерисовать виджет с текущими данными
 }