+++ src/interfaces/graphQL/aggregationDataSource.ts
@@ -3,12 +3,12 @@
  * Используется для последующей комбинации двух вложенных источников данных.
  * result = firstDataSource {operation} secondDataSource
  */
-import {DataSourceType, Operation} from '../../models/types';
+import {ArithmeticOperation, DataSourceType} from '../../models/types';
 import {DataSource} from "./dataSource";
 
 export interface AggregationDataSource {
     type: DataSourceType;
     firstDataSource: DataSource;        // первый вложенный источник данных
-    operation: Operation;               // операция между первым и вторым источником данных.
+    operation: ArithmeticOperation;     // операция между первым и вторым источником данных.
     secondDataSource: DataSource;       // второй вложенный источник данных.
 }