+++ src/models/types.ts
@@ -1,29 +1,20 @@
-export interface Paddings {
-    top: number;
-    right: number;
-    bottom: number;
-    left: number;
-}
+export * from "./typesGraphQL";
 
 // Типы виджетов
 // Каждый тип виджета может поддерживать определенное число различный viewType
 export type WidgetType = 'SPLINE' | 'AVERAGE_NUMBER' | 'SOLID_GAUGE' | 'INDICATORS_TABLE' | 'TABLE'
      | 'REPORT' | 'STATIC' | 'KPI' | 'AVATAR' | 'DISTRIBUTION' | 'PROFILE';
 
-// Типы отображения виджетов. Именно от них формируются запросы в graphQL.
-export type ViewType = 'STATIC' | 'DYNAMIC' | 'REPORT' | 'TABLE' | 'DISTRIBUTION' | 'PROFILE'; // | 'MAP'
-
 export type ChartType = 'LINE' | 'HISTOGRAM' | 'SCATTER' | 'COMPARED_FACT' | 'COMPARED_PLAN';
 
+// Стиль линии графиков
+export const LineTypeValues = ['solid', 'dashed', 'dotted'] as const;
+export type LineType = typeof LineTypeValues[number];
+
+// Позиция оси Y
 // NOTE: Нельзя через enum, т.к. они не экспортируются, поэтому через массив значений + type
-export const YAxisTypesValues = ['left', 'right'] as const;   // Создаем массив допустимых значений
+export const YAxisTypesValues = ['left', 'right'] as const;   // Создаем массив допустимых значений для экспорта в конфиге
 export type YAxisTypes = typeof YAxisTypesValues[number];
 
 export type MethodType = 'MAPE' | 'MAE' | 'coverage' | 'utilization';
 export type ServerType = 'druid' | 'qlik';
-
-// GraphQL enums
-export type Frequency = 'YEAR' | 'MONTH' | 'WEEK' | 'DAY' | 'HOUR' | 'ALL';
-export type Operation = 'SUM' | 'AVG' | 'MIN' | 'MAX';
-export type DataSourceType = 'SINGLE' | 'AGGREGATION';
-export type ArithmeticOperation = 'SUM' | 'SUBTRACT' | 'MULTIPLY' | 'DIVIDE';