+++ src/widgetSettings/index.ts
@@ -1,0 +1,2 @@
+export * from './IWidgetSettings';
+export * from './makeSettings';