+++ src/widgets/avatar/settings.ts
@@ -1,0 +1,16 @@
+import {IWidgetSettings} from "../../interfaces";
+import {makeSettings} from "../../widgetSettings";
+import {makeColor, makeString} from "../../widgetSettings/settings";
+
+export const settings: IWidgetSettings = makeSettings({
+    settings: [
+        makeString('title', 'Заголовок', ''),
+    ],
+    dataSet: {
+        initAmount: 1,
+        canAdd: true,
+        settings: [
+            makeColor('color', ' Цвет', null),
+        ]
+    }
+});