+++ src/widgets/indicatorsTable/index.ts
@@ -1,2 +1,2 @@
 export * from "./indicatorsTable";
-export * from "./config";
+export * from "./settings";