+++ src/widgets/solidGauge/solidGauge.ts
@@ -1,8 +1,8 @@
 import s from "../../styles/_all.less";
 import w from "./solidGauge.less";
-import {config as widgetConfig} from "./config";
+import {settings as widgetSettings} from "./settings";
 
-import {IChartData, IWidgetVariables} from "../../interfaces";
+import {IChartData, IWidgetSettings, IWidgetVariables} from "../../interfaces";
 import {get as _get} from "lodash";
 import {Chart} from "../../models/Chart";
 import {TypeGuardsHelper} from "../../helpers";
@@ -12,13 +12,19 @@ export class SolidGauge extends Chart {
         return {};
     }
 
-    run(data: IChartData): void {
+    getSettings(): IWidgetSettings {
+        return widgetSettings;
+    }
+
+    run(): void {
+        const data: IChartData = this.chartData;
+
         if (TypeGuardsHelper.dataSetsIsDataSetTemplate(data.dataSets)) {
             const maxValue = _get(data, 'data[1][0].value', 0);
             const currValue = _get(data, 'data[0][0].value', 0);
 
-            const color = this.getColor(widgetConfig, data.dataSets[0].settings, 'color-yellow');
-            const maxColor = this.getColor(widgetConfig, data.dataSets[1].settings, 'color-grey');
+            const color = this.getColor(data.dataSets[0].settings, 'color-yellow');
+            const maxColor = this.getColor(data.dataSets[1].settings, 'color-grey');
 
             const percent = currValue / maxValue * 100;
             const magicLengthOfSvgPath = 503.3096923828125;
@@ -29,8 +35,8 @@ export class SolidGauge extends Chart {
             const output = this.renderTemplate({
                 fontSize,
                 currValue,
-                title: this.getWidgetSetting(widgetConfig, data.settings, 'title'),
-                icon: this.getWidgetSetting(widgetConfig, data.settings, 'icon'),
+                title: this.getWidgetSetting(data.settings, 'title'),
+                icon: this.getWidgetSetting(data.settings, 'icon'),
                 maxColor: maxColor.color,
                 color: color.color,
                 magicLengthOfSvgPath,